refacto(type-designer): split card component in multiple component ro enhance readability
